Add Rails asset pipeline support.
